refactor(组件结构): 将数据刷新功能从Home移动到HelloSpringWorld组件
重构组件结构，将数据刷新逻辑和状态管理从Home组件移至HelloSpringWorld组件，使功能职责更清晰
